src/log/slog: JSONHandler formats time.Duration as string
Change JSONHandler to format durations with Duration.String.

For the most part, JSONHandler conforms to encoding/json.Marshal,
which formats time.Duration as an integer number of nanoseconds.
Using the output of time.Duration.String is better for two reasons:

- No unit is associated with the integer. You have to know it's nanos.
  If you didn't already know it was generated by Go, you'd probably
  think it was seconds. The string format's interpretation is clear.

- The integer can overflow the maximum representable JSON integer.
  encoding/json will correctly switch to a string in that case,
  but naive JSON parsers may find that confusing.

This change will slow down JSONHandler for durations, but favoring
readability over speed is the right choice in this case. One can
always write a ReplaceAttr function to return an int64 and get
the fast behavior.

Change-Id: Id49e3ff43819f162bf93a6d0361d0d09e497b11f