runtime: add Go directive "go:nocheckptr" to forkAndExecInChild1
forkAndExecInChild1 function must not acquire any locks in child, because they might have been locked at the time of the fork. This implies to no rescheduling, no malloc calls, and no new stack segments.

So, doing a "checkptrAlignment" is bad here, because checkptr functions added by the instrumentation could grow the stack, which we should not be done between fork and exec.

Hence using a Go directive "go:nocheckptr" to forkAndExecInChild1() func, so that the compiler should not do "checkptrAlignment" when functions marked with "go:norace".

This race detection bug was caught in go 1.21 on s390x.
Running a "./race.bash" script from "go/src" directory failed and the bug details are provided in issue link mentioned below.

This CL will fix following bug:
https://github.com/golang/go/issues/58785

Change-Id: I254091368b0789d886acdf26f8aa8d8f5a986b24