sync: Do not keep the function passed to OnceXxx alive after the first call.
The function passed to OnceFunc/OnceValue/OnceValues may transitively keep
more allocations alive. As the passed function is guaranteed to be called
at most once, it is safe to drop it after the first call is complete. This
avoids keeping the passed function (and anything it transitively references)
alive until the returned function is GCed.

Change-Id: I2faf397b481d2f693ab3aea8e2981b02adbc7a21