MIPS: add mips30r2/mips32r5 and FPXX support
Support options like "mips32r2,hardfloat" for mips.
The possible combinations are:
	[softfloat|hardfloat|],[mips32,mips32r2,mips32r5,]
The hardfloat and mips32 are the default, and thus can be ignored.

When hardfloat,mips32(r1) is used, the FPU ABI is set to FP32.
In this case, to load a float64, we use:
	lwc1 $fEVEN,IMM(GPR)
	nop
	lwc1 $F(EVEN+1),(IMM+4)(GPR)
	nop

When hardfloat,mips32r2/mips32r5 is used, the FPU ABI is set to FPXX.
In this case, to load a float64, we use:
	lwc1 $fEVEN,IMM(GPR)
	nop
	lw   REGTMP,(IMM+4)(GPR)
	mthc1 REGTMP, $FEVEN

There are some advantages of FPXX:
	1.On some hardware, FP32 may have very bad performance.
	2.The default value of FPU ABI of GCC and LLVM is FPXX.
	3.FPXX objects can interlink with both FP32 and FP64.
	4.FPXX executable can run on both FR=1 and FR=0 mode.

For more information:
https://web.archive.org/web/20180828210612/https://dmz-portal.mips.com/wiki/MIPS_O32_ABI_-_FR0_and_FR1_Interlinking

Change-Id: I56508d7082dc7974a9033b03985c87e57d6e8ec5