runtime: zero saved frame pointer when reusing goroutine stack on ARM
When a goroutine stack is reused on ARM, the spot on the stack where the
"caller's" frame pointer goes for the topmost frame should be explicitly
zeroed. Otherwise, the frame pointer check in adjustframe with
debugCheckBP enabled will fail on the topmost frame of a call stack the
first time a reused stack is grown.

Updates #39524, #58432

Change-Id: Ic1210dc005e3ecdbf9cd5d7b98846566e56df8f5