cmd/go: return and handle errors from loadImport for bad imports
And assign the error to the importing package. Before this change, for
some errors for bad imports, such as importing a vendored package
with the wrong path, we would make a dummy package for the imported
package with the error on it. Instead report the error on the
importing package where it belongs. Do so by returning an error from
loadImport and handling it on the importing package.

For #59157

Change-Id: I952e1a82af3857efc5da4fd3f8bc6be473a60cf5