go/types, types2: implement reverse type inference for function arguments
Allow function-typed function arguments to be generic and collect
their type parameters together with the callee's type parameters
(if any). Use a single inference step to infer the type arguments
for all type parameters simultaneously.

Requires Go 1.21 and that Config.EnableReverseTypeInference is set.
Does not yet support partially instantiated generic function arguments.
Not yet enabled in the compiler.

For #59338.

Change-Id: Ia1faa27a28c6353f0bbfd7f81feafc21bd36652c