go/types, types2: generalize tparamIndex to arbitrary type parameter lists
tparamIndex returns the index of a type parameter given the type
parameter and a list of type parameters. For it to work correctly,
the type parameter list must be of a single type parameter declaration,
so that the respective type parameter indices match the type parameter
list entry indices. This allows for fast lookup.

To allow for lists of arbitrary type parameters (from different generic
signatures), change the implementation to do a linear search. This is
likely good enough because the total number of type parameters tends
to be small (< 10).

Change-Id: I913f97fa4c042abeb535ee86ca6657241a4cf796