cmd/asm: do not pass scale to backend of assembler in arm64
In Arm64, we accept scale 0/1 for backend compatibility in CL 328229,
the scale is recorded in Addr.Scale but never used when we're encoding
the instructions, which will be a little bit waste.

This CL absorbs the scale in the parser so the Addr.Scale is always
empty in Arm64, then we can take it for other use in register encoding.

Change-Id: I3cf9a0610069242b2268746cce7ba14aba802098